Refactor layout of this project.
This changes the repo to contain a single "loadsjs" project with custom
test runner.  It drops the mocha-reporter support and implements an
experimental new interface between loads and loadsjs, which should help
support generation of higher load.
