+++ .gitignore
@@ -1,0 +1,3 @@
+*~
+*.swp
+/node_modules
