--- loads.js/README.md
@@ -1,19 +1,0 @@
-# Loads.js
-
-A library to send data to the [loads](http://loads.rtfd.org) runner.  This
-library doesn't do the integration with the test frameworks you want to use.
-Rather, it provides a high level interface you can work with.
-
-If you are looking for an integration with mocha, for instance, have a look
-[here](https://github.com/mozilla-services/loads.js/mocha-loads).
-
-## Installation
-
-You can install loads.js with npm. Go in your project and run
-
-```
-$ npm install ../../services/loads.js
-```
-
-## Usage
-
