--- loads.js/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "loads.js",
-  "version": "0.0.1",
-  "description": "A reporter to send the results to the loads runner",
-  "main": "index.js",
-  "repository": "http://github.com/mozilla-services/loads.js",
-  "license": "MPL 2.0",
-  "dependencies": {
-    "zmq": "2.4.x"
-  }
-}
