--- mocha-loads/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "mocha-loads-reporter",
-  "version": "0.0.1",
-  "description": "A reporter to send the results to the loads runner",
-  "main": "index.js",
-  "directories": {
-  },
-  "scripts": {
-  },
-  "repository": "http://github.com/mozilla-services/loads.js",
-  "license": "MPL 2.0",
-  "dependencies": {
-    "loads.js": "0.0.1"
-  },
-  "devDependencies": {
-    "mocha": "1.9.x"
-  }
-}
